updating the delete endpoint

+++ API/url_shortner/url_short.py
@@ -6,7 +6,7 @@ import random
 from ..model import Url_link,User
 from flask_jwt_extended import jwt_required,get_jwt_identity
 from sqlalchemy import and_
-from flask import redirect,send_file,request
+from flask import redirect,send_file,request,g
 import qrcode
 from ..utils import cache,db
 import os
@@ -61,21 +61,21 @@ class Url_short(Resource):
      @url_namespace.marshal_with(url_short)
      @url_namespace.response(200, url_short)
      @jwt_required()
-     @cache.cached(timeout=3600,key_prefix=lambda: f"user_links:{get_jwt_identity()}")
+     
      @url_namespace.response(400, 'Bad Request')
      def get(self):
         '''get all the links shortened by a particular user'''
         user_id=get_jwt_identity()
         page = request.args.get('page', 1, type=int)
         per_page = request.args.get('per_page', 10, type=int)
-        cache_key = f"user_links:{user_id}:{page}:{per_page}"
-        cached_result = cache.get(cache_key)
-        if cached_result is not None:
-            return cached_result
+        
+        
+
         links=Url_link.query.filter(Url_link.user_id==user_id).order_by(Url_link.date_posted.desc()).paginate(page=page, per_page=per_page)
         response_headers = {'X-Total-Count': links.total}
         result = links.items, 200, response_headers
-        cache.set(cache_key, result)
+        
+       
         return result
 
 
@@ -88,16 +88,17 @@ class Url_shortDel(Resource):
     @jwt_required()
     def delete(self,id):
         """delete the short url"""
-        url=Url_link.query.get_or_404(id)
-        url.delete()
-        return {"status":"success","message":"url has been successfully deleted"},200
-    
+        user_id=get_jwt_identity()
+        if Url_link.query.filter(Url_link.user_id==user_id).first():
+            url=Url_link.query.get_or_404(id)
+            url.delete()
+            return {"status":"success","message":"url has been successfully deleted"},200
+        abort(403,message="you can delete this url, it isnt owned by you",status="error")
+        
 
 @url_namespace.route("/<short_url>")
 class Url_Shortened(Resource):
     @cache.memoize(timeout=3600)
-    @jwt_required()
-    @url_namespace.doc(security="jwt")
     @url_namespace.response(200, "url redirect successful")
     @url_namespace.response(400, 'Bad Request')
     def get(self,short_url):
